requirements.txt
google-cloud-bigquery
rtmbot
oauth2client